requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
uvicorn[standard]==0.24.0.post1  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==5.0  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.59.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-tailwind==0.5  # https://github.com/django-crispy-forms/crispy-tailwind
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-webpack-loader==2.0.1  # https://github.com/django-webpack/django-webpack-loader


# Project
# ------------------------------------------------------------------------------
django-imagekit
django-phonenumber-field[phonenumbers]


# API
# ------------------------------------------------------------------------------
djangorestframework
django-cors-headers
